backend/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
SQLAlchemy>=2.0.0
alembic>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic-settings>=2.0.0

# Security and scanning
zaproxy>=0.4.0
supabase>=2.0.0
wapiti3>=3.2.0
mitmproxy>=11.0.0
python-owasp-zap-v2.4>=0.4.2
nuclei-sdk>=0.1.0
python-nuclei>=0.2.0
aiohttp-client-cache>=0.8.1
PyJWT>=2.6.0
cryptography>=40.0.0

# ML and data processing
numpy>=1.20.0
pandas>=2.0.0
scikit-learn>=1.0.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
torch>=2.0.0
shap>=0.42.0
mlflow>=2.8.0
optuna>=3.3.0
feast>=0.31.0

# Utilities and monitoring
redis>=6.0.0
loguru>=0.7.0
tqdm>=4.60.0

# Web and networking
aiohttp>=3.8.0
requests>=2.20.0
httpx>=0.20.0
websockets>=11.0.0
fastapi-websocket-pubsub>=0.3.0
asyncio>=3.4.3

# Development and testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
python-dotenv>=1.0.0